+++ pong_front/src/assets/Pingu3D.png
